requirements.txt
django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-authtoken>=1.3,<1.4

